Implement a `isReturnValueUsed` utility
