+++ rules/prefer-node-append.js
@@ -1,40 +1,18 @@
 'use strict';
 const getDocumentationUrl = require('./utils/get-documentation-url');
+const isReturnValueUsed = require('./utils/is-return-value-used');
 
 const getMethodName = memberExpression => memberExpression.property.name;
-
-const ignoredParentTypes = [
-	'ArrayExpression',
-	'IfStatement',
-	'MemberExpression',
-	'Property',
-	'ReturnStatement',
-	'VariableDeclarator'
-];
-
-const ignoredGrandparentTypes = [
-	'ExpressionStatement'
-];
 
 const create = context => {
 	return {
 		CallExpression(node) {
-			const {
-				callee,
-				parent
-			} = node;
+			const {callee} = node;
 
 			if (callee.type === 'MemberExpression' && getMethodName(callee) === 'appendChild') {
 				let fix = fixer => fixer.replaceText(callee.property, 'append');
 
-				const {
-					parent: grandparent
-				} = (parent || {});
-
-				if (
-					(parent && ignoredParentTypes.includes(parent.type)) ||
-					(grandparent && ignoredGrandparentTypes.includes(grandparent.type))
-				) {
+				if (!isReturnValueUsed(node)) {
 					fix = undefined;
 				}
