+++ rules/utils/is-return-value-used.js
@@ -1,0 +1,24 @@
+'use strict';
+
+const ignoredParentTypes = [
+	'ArrayExpression',
+	'IfStatement',
+	'MemberExpression',
+	'Property',
+	'ReturnStatement',
+	'VariableDeclarator'
+];
+
+const ignoredGrandparentTypes = [
+	'ExpressionStatement'
+];
+
+module.exports = function (node) {
+	const {parent} = node;
+	const {
+		parent: grandparent
+	} = (parent || {});
+
+	return (!parent || !ignoredParentTypes.includes(parent.type)) &&
+		(!grandparent || !ignoredGrandparentTypes.includes(grandparent.type));
+};
